Fix NJ details leaking into other brands' quotations; allow decimal quantities
Two customer-visible bugs in the quotation document.

1. NJ company details on the wrong document / missing on the right one.
   companyProfileForBrand() lost its `data.company` fallback in 7886810 (an
   unrelated mobile-PDF commit): it ignored its `data` argument entirely and
   never returned `isGlobalFallback`, which the document reads. So the seeded
   "nj" brand - which carries a name and nothing else - printed a bare "NJ"
   header with NO address and NO contact row, and a quotation with no
   resolvable brand printed a completely blank company block.
   Restored the per-field fallback the Brands settings screen already promises
   ("Empty fields fall back to the original NJ company details"), with the
   inheritance still gated to the NJ brand - letting any other brand inherit is
   exactly what would print NJ's address and phone on a Highlander quotation.
   For NJ the company name leads over the brand name, since `brand.name` is the
   short label ("NJ") while the header must carry the legal entity.

2. Decimal quantities and prices were impossible to type on the document.
   EditableCell rendered a CONTROLLED <input type="number">. The moment the
   user pressed ".", the element's value was "7.", which is not a valid
   floating-point number per the HTML spec, so the browser sanitised
   `e.target.value` to "" and wiped the digits already typed. Quantities are
   measured in sqft/RFT and are routinely fractional (7.5), and prices carry
   paise (1250.50), so neither could be entered by clicking the cell.
   Numeric cells are now type="text" + inputMode="decimal" (which also gives
   Android a keypad WITH a decimal key) and filter their own keystrokes - the
   same approach NumberField already used for the cart/desk fields.

The keystroke filter was duplicated between the two, so it moves to
frontend/src/numeric.js as sanitizeDecimal() and both import it. Being plain
JS rather than JSX, it is directly exercisable by node.

Tests: numeric.test.mjs covers every intermediate state of typing "7.5" plus
the multi-dot, junk-input and negative cases; brands.test.mjs gains assertions
that a non-NJ brand's profile contains no NJ data at all.

Version bumped to 1.1.1 so installed 1.1.0 clients are offered the update.

+++ frontend/src/components/Checkout.jsx
@@ -470,7 +470,9 @@ export default function Checkout() {
   };
 
   const handleQtyChange = (cartId, newQty) => {
-    setCart(prev => prev.map(item => item.cartId === cartId ? { ...item, qty: Math.max(1, parseInt(newQty) || 1) } : item));
+    // parseFloat, not parseInt: quantities are measured in sqft/RFT and are
+    // routinely fractional (7.5 sqft). parseInt silently truncated them.
+    setCart(prev => prev.map(item => item.cartId === cartId ? { ...item, qty: Math.max(1, parseFloat(newQty) || 1) } : item));
   };
 
   const handleUnitChange = (cartId, newUnit) => {
